Add Support for typing.NewType (#32)
typing.NewType is actual just a function and should be interpreted as
an alias for something. However its __str__
‘NewType.<locals>.new_type’, so it includes neither its custom name nor
the type for which it is an alias, which is not very helpful to
identify why a certain type check failed. Hence, such information
are also printed in error messages, but other types are left as they
were before.

Signed-off-by: Fabian Raab <fabian@raab.link>
Signed-off-by: Wilfred Wong <keiclone@users.noreply.github.com>